Updated Java version

+++ Codechef/Guess/Guess.java
@@ -13,15 +13,15 @@ class Guess
         int T = sc.nextInt();
         while (T-- > 0)
         {
-            int N = sc.nextInt();
-            int M = sc.nextInt();
+            long N = sc.nextInt();
+            long M = sc.nextInt();
             long total = N * M;
             if (N == 1 && M == 1)
                 System.out.println("0/1");
 
             else if (N % 2 == 0 || M % 2 == 0)
             {
-                System.out.println("1/2" );
+                System.out.println("1/2");
             }
 
             else
